Replace the Poetry-based dependency handling with a Dockerfile.
As it turns out, putting together a compatible execution environment can
be exceedingly difficult on distributions other than Arch or Gentoo.
Switching to a Docker-based solution ensures the correct dependencies
are available.

A current weakness is that dependencies are not pinned in any way, and
as Arch (the distribution used by the Dockerfile) can be a fast-moving
target, it's possible errors will come up sooner rather than later.

Signed-off-by: Jason Lynch <jason@aexoden.com>