Handle some GPS log edge cases with this dashcam model.
Normally, the dashcam fairly reliably outputs one GPS sentence per
second, and the values actually increase monotonically and consistently.
Occasionally, however, the GPS update sequence will be out of phase and
result in it missing some seconds and outputting other seconds twice.

Previously, the code just assumed each sentence was for the next second.
Now, it instead checks the timestamps on the sentences and inserts empty
entries as necessary that will ultimately be filled in by the existing
interpolation code. Duplicate entries are also removed. This results in
a smoother map video when this happens.

Signed-off-by: Jason Lynch <jason@aexoden.com>